requirements_mac.txt
# Core dependencies for Mac M3
unsloth @ git+https://github.com/unslothai/unsloth.git
torch>=2.2.0
transformers>=4.37.2
peft>=0.7.1
accelerate>=0.25.0
trl>=0.7.4
datasets>=2.16.1
huggingface_hub>=0.20.3

# Utilities
tqdm>=4.66.1
packaging
python-dotenv
matplotlib
seaborn
pandas

# Optional: For API serving if needed
fastapi
uvicorn
pydantic
requests 